First pass towards getting HCatalog artifacts in maven central.
This change adds a new "mvn-install" ant task that installs artifacts in the local maven repo. However, it does not work correctly at this time because we do not generate correct pom files for hcatalog-core or hcatalog-server-extensions. This causes them to be published as "hcatalog" which is incorrect.

Next step is generating correct pom files.